# CRUD_review/consult-review/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para obtener reseñas de lugares turísticos.",
        "title": "Microservicio de Consulta de Reseñas",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/reviews": {
            "get": {
                "description": "Devuelve un listado de todas las reseñas almacenadas en la base de datos.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Obtiene todas las reseñas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Review"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reviews/place/{place_id}": {
            "get": {
                "description": "Devuelve todas las reseñas asociadas a un lugar, según el ID del lugar.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Obtiene reseñas de un lugar turístico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del lugar turístico",
                        "name": "place_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Review"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reviews/{id}": {
            "get": {
                "description": "Devuelve una reseña según el ID proporcionado en la URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Obtiene una reseña específica",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la reseña",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Review"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Review": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "place_name": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}